feat: 優化啟動通知並添加權益總值顯示
主要更新:
- 移除 KBar.volume 欄位（期貨不需要）
- 將啟動通知從 main.py 移至 TradingService.run_strategy
- 優化 send_status_message 支持參數化消息生成
- 添加權益總值 (Total Equity) 顯示
- 調整消息格式，移除多餘空格對齊
- 啟動時自動查詢並顯示賬戶權益總值

技術細節:
- LineBotService: send_status_message 接收獨立參數
- TradingService: 調用 account_service.get_margin() 獲取權益
- 消息格式簡潔，去除過多空格對齊
- 向後兼容舊的 status 參數用法

+++ src/auto_trade/models/market.py
@@ -342,8 +342,8 @@ class Quote:
     """即時報價資料模型"""
 
     symbol: str
-    price: float
+    price: int
     volume: int
-    bid_price: float | None = None
-    ask_price: float | None = None
+    bid_price: int | None = None
+    ask_price: int | None = None
     timestamp: datetime | None = None
